backend/requirements.txt
flask
pymongo
authlib
flask-cors
flask-mail
pyjwt
twilio
werkzeug
python-dotenv
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
opencv-python-headless==4.10.0.84
numpy<2.0

Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
torch>=2.0.0
torchvision
opencv-python-headless
numpy
pillow
gunicorn
huggingface_hub==0.23.4
transformers==4.30.2